Added 2 more modules, plus a modloader
Should be the final commit for β 0.7 .
It doesn't have the amount of modules I was originally planning to add but it's nearly been a full year since I updated so I might as well.

New modules:
Alchemy
Binary LEDs

New additions:
Modloader (Loads custom .class files)

Also on a sidenote: This is the first commit made on JetBrains IDE so expect some merging issues.